backend/requirements.txt
fastapi[standard]
python-socketio
anthropic
networkx
pydantic
python-dotenv
